DSND-1536: Finish integration tests *modified pom to remove unnecessary maven-compiler-plugin *refactored steps files for iTests
